Gelatin v1.1 results
The modified preparation process made for baits that were clear and free of air bubbles, but the elastic modulus of the final product was too low to make these baits viable. I suspect that the sugar wasn't cooked at a high enough temperature to cause sufficient inversion. Additional experimentation along this line are warranted.